feat: authentication + dev database included in docker compose

+++ app/core/config.py
@@ -17,7 +17,7 @@ class AdminSettings(BaseSettings):
 class JWTSettings(BaseSettings):
     jwt_secret_key: str
     jwt_algorithm: str = "HS256"
-    access_token_expire_minutes: int = 30
+    access_token_expire_minutes: float = 30
 
     model_config = SettingsConfigDict(env_file=".env", extra="ignore")
